+++ CHANGELOG.md
@@ -1,6 +1,5 @@
 # Change Log
 ---
 
-- [Add metadata transfer example.](https://github.com/Tencent/spring-cloud-tencent/pull/211)
-- [feat:merge features from 1.5.x-Hoxton.SR9.](https://github.com/Tencent/spring-cloud-tencent/pull/250)
 - [Upgrade: fix third-party lib CVEs & upgrade core spring libs version](https://github.com/Tencent/spring-cloud-tencent/pull/263)
+- [feat:support reading configuration from application.yml or application.properties.](https://github.com/Tencent/spring-cloud-tencent/pull/262)